Lift single-firing replay out of HarnessRunner
fire_extractor_from_record / fire_auditor_from_record never used the
runner's cumulative state, child runner, sink, sidecar, or intervals —
they rebuilt inputs from the record and called run_phase_standalone
directly. So replay/runner.py constructed a whole HarnessRunner (with
NoopSink, AuditorSettings.empty(), a fresh cumulative state) purely as
a costume to host one method call.

- Move both bodies into standalone functions in replay/runner.py;
  public replay_extractor_record / replay_auditor_record signatures are
  unchanged, so chain.py / cli.py / rl_prompts are unaffected.
- Delete fire_*_from_record + _coerce_schema_list from the runner, plus
  the now-dead cwd constructor param/field and the orphaned
  run_phase_standalone / PhaseResult / build_extractor_directive imports.
- Remove AuditorSettings.empty() (its sole purpose was satisfying the
  runner constructor on the replay path) and NoopSink (no remaining
  consumer).
- Trim runner __all__ to the two helpers the seams/adapter re-export.
- Refresh docstrings to the current shape: drop the P1/P2/P3 phase
  framing and the "lifted verbatim from legacy _x" references; convert
  the firing_id-on-restart TODO into an honest known-limitation note.

Behavior-preserving (the moved code is verbatim). Validated: 89 passed
+ 1 pre-existing xfail, mypy strict clean, import smoke (no circular
import), and a real-provider extractor replay over a recorded sidecar
produced a valid witness-bearing graph.
